+++ gunicorn/__main__.py
@@ -2,7 +2,6 @@
 #
 # This file is part of gunicorn released under the MIT license.
 # See the NOTICE for more information.
-import http
-print(dir(http))
+
 from gunicorn.app.wsgiapp import run
 run()
