+++ tests/requests/invalid/009.py
@@ -1,2 +1,2 @@
-from gunicorn.ghttp.errors import LimitRequestHeaders
+from gunicorn.http.errors import LimitRequestHeaders
 request = LimitRequestHeaders
