+++ tests/requests/invalid/016.py
@@ -1,3 +1,3 @@
-from gunicorn.ghttp.errors import InvalidRequestLine
+from gunicorn.http.errors import InvalidRequestLine
 
 request = InvalidRequestLine
