+++ tests/requests/invalid/019.py
@@ -1,5 +1,5 @@
 from gunicorn.config import Config
-from gunicorn.ghttp.errors import InvalidSchemeHeaders
+from gunicorn.http.errors import InvalidSchemeHeaders
 
 request = InvalidSchemeHeaders
 cfg = Config()
