+++ tests/requests/invalid/pp_01.py
@@ -1,5 +1,5 @@
 from gunicorn.config import Config
-from gunicorn.ghttp.errors import InvalidProxyLine
+from gunicorn.http.errors import InvalidProxyLine
 
 cfg = Config()
 cfg.set("proxy_protocol", True)
