+++ tests/test_http.py
@@ -6,10 +6,10 @@ import pytest
 import unittest.mock as mock
 
 from gunicorn import util
-from gunicorn.ghttp.body import Body, LengthReader, EOFReader
-from gunicorn.ghttp.wsgi import Response
-from gunicorn.ghttp.unreader import Unreader, IterUnreader, SocketUnreader
-from gunicorn.ghttp.errors import InvalidHeader, InvalidHeaderName
+from gunicorn.http.body import Body, LengthReader, EOFReader
+from gunicorn.http.wsgi import Response
+from gunicorn.http.unreader import Unreader, IterUnreader, SocketUnreader
+from gunicorn.http.errors import InvalidHeader, InvalidHeaderName
 
 
 def assert_readline(payload, size, expected):
